+++ R/plot_proportions.R
@@ -44,6 +44,7 @@
 #' @param overall_line_width Line width of the overall line.
 #' @param facet_labeller Labeller function for strip labels.
 #' @param flip Flip x and y axis?
+#' @param free_scale Allow y axis to vary between conditions?
 #' @param return_data Return data used instead of the plot?
 #' @export
 #' @keywords univar
@@ -113,6 +114,28 @@
 #'  )
 #'
 #' # defining several proportions
+#'
+#' titanic |>
+#'   plot_proportions(
+#'     dplyr::tibble(
+#'       Survived = Survived == "Yes",
+#'       Male = Sex == "Male"
+#'     ),
+#'     by = c(Class),
+#'     mapping = ggplot2::aes(fill = condition)
+#'   )
+#'
+#' titanic |>
+#'   plot_proportions(
+#'     dplyr::tibble(
+#'       Survived = Survived == "Yes",
+#'       Male = Sex == "Male"
+#'     ),
+#'     by = c(Class),
+#'     mapping = ggplot2::aes(fill = condition),
+#'     free_scale = TRUE
+#'   )
+#'
 #' iris |>
 #'   plot_proportions(
 #'     dplyr::tibble(
@@ -185,6 +208,7 @@ plot_proportions <- function(
   overall_line_width = .5,
   facet_labeller = ggplot2::label_wrap_gen(width = 50, multi_line = TRUE),
   flip = FALSE,
+  free_scale = FALSE,
   return_data = FALSE
 ) {
   # variable identification
@@ -392,9 +416,17 @@ plot_proportions <- function(
             .data$condition
           ) |>
           dplyr::summarise(num_level = dplyr::last(.data$num_level)),
-        by = "variable"
+        by = c("variable", "condition")
+      ) |>
+      dplyr::left_join(
+        d |>
+          dplyr::group_by(.data$condition) |>
+          dplyr::summarise(
+            y = ifelse(show_ci, max(.data$prop_high), max(.data$prop))
+          ),
+        by = "condition"
       )
-    pvalues$y <- ifelse(show_ci, max(d$prop_high), max(d$prop))
+    if (!free_scale) pvalues$y <- max(pvalues$y)
     pvalues$label <- pvalues_labeller(pvalues$p)
 
     plot <-
@@ -463,7 +495,10 @@ plot_proportions <- function(
   plot <-
     plot +
     ggplot2::labs(x = NULL, y = NULL) +
-    ggplot2::scale_y_continuous(labels = scales::percent) +
+    ggplot2::scale_y_continuous(
+      labels = scales::percent,
+      expand = ggplot2::expansion(mult = c(0, .1))
+    ) +
     ggplot2::scale_x_discrete(
       labels = \(x) {
         stringr::str_remove(x, "^[0-9]*_")
@@ -485,7 +520,7 @@ plot_proportions <- function(
       ggplot2::facet_grid(
         rows = ggplot2::vars(.data$variable_label),
         cols = cond_facet,
-        scales = "free_y",
+        scales = ifelse(free_scale, "free", "free_y"),
         space = "free_y",
         labeller = facet_labeller,
         switch = "y"
@@ -506,7 +541,7 @@ plot_proportions <- function(
       ggplot2::facet_grid(
         cols = ggplot2::vars(.data$variable_label),
         rows = cond_facet,
-        scales = "free_x",
+        scales = ifelse(free_scale, "free", "free_x"),
         space = "free_x",
         labeller = facet_labeller
       ) +